tests/wpt/tests/css/css-gaps/parsing/rule-interior-inset-bidirectional-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Gap Decorations: rule-interior-inset shorthands</title>
<link rel="help" href="https://www.w3.org/TR/css-gaps-1/#inset">
<meta name="assert"
    content="rule-interior-inset supports the full grammar '<length-percentage> <length-percentage>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
    <script>
        const rule_properties = {
            'rule-interior-inset': ['column-rule-interior-inset-start', 'column-rule-interior-inset-end',
                'row-rule-interior-inset-start', 'row-rule-interior-inset-end']
        };

        const bidirectionalTestCases = [
            {
                input: '0px',
                expected: {
                    columnRuleInteriorInsetStart: '0px',
                    columnRuleInteriorInsetEnd: '0px',
                    rowRuleInteriorInsetStart: '0px',
                    rowRuleInteriorInsetEnd: '0px'
                }
            },
            {
                input: '10px',
                expected: {
                    columnRuleInteriorInsetStart: '10px',
                    columnRuleInteriorInsetEnd: '10px',
                    rowRuleInteriorInsetStart: '10px',
                    rowRuleInteriorInsetEnd: '10px'
                }
            },
            {
                input: '50%',
                expected: {
                    columnRuleInteriorInsetStart: '50%',
                    columnRuleInteriorInsetEnd: '50%',
                    rowRuleInteriorInsetStart: '50%',
                    rowRuleInteriorInsetEnd: '50%',
                }
            },
            {
                input: '10px 20px',
                expected: {
                    columnRuleEdgeInsetStart: '10px',
                    columnRuleEdgeInsetEnd: '20px',
                    columnRuleInteriorInsetStart: '10px',
                    columnRuleInteriorInsetEnd: '20px',
                    rowRuleEdgeInsetStart: '10px',
                    rowRuleEdgeInsetEnd: '20px',
                    rowRuleInteriorInsetStart: '10px',
                    rowRuleInteriorInsetEnd: '20px',
                }
            },
        ];


        for (rule_property in rule_properties) {
            const test_cases = bidirectionalTestCases;

            for (const { input, expected } of test_cases) {
                const [columnRuleInteriorInsetStart, columnRuleInteriorInsetEnd, rowRuleInteriorInsetStart, rowRuleInteriorInsetEnd] = rule_properties[rule_property];

                test_shorthand_value(rule_property, input, {
                    [columnRuleInteriorInsetStart]: expected.columnRuleInteriorInsetStart,
                    [columnRuleInteriorInsetEnd]: expected.columnRuleInteriorInsetEnd,
                    [rowRuleInteriorInsetStart]: expected.rowRuleInteriorInsetStart,
                    [rowRuleInteriorInsetEnd]: expected.rowRuleInteriorInsetEnd,
                });
            }
        }
    </script>
</body>
</html>
